+++ PatternAnalysis/activeLearning/mclr_SM.h
@@ -55,7 +55,7 @@ public:
 	vnl_matrix<double> train_data;
 	vnl_vector<double> y; // labels (-1 for training)	
 	vnl_vector<double> y_ground_truth;	
-	std::vector< std::pair<double,double> > id_time_val;	
+	std::vector< std::pair<int,int> > id_time_val;	
 	vnl_matrix<double> z; // used in gradient computation	
 	vnl_matrix<double> gradient_w; // used in gradient computation	 
 	vnl_matrix<double> hessian;
@@ -113,7 +113,7 @@ public:
 	vnl_matrix<double> Kron(vnl_vector<double> x,vnl_vector<double> y);
 	void Get_Label_Sample(int query);
 	FILE* FDeclare2(char *root, char *extension, char key);
-	vnl_matrix <double> tableToMatrix(vtkSmartPointer<vtkTable> table,std::vector< std::pair<double,double> > id_list);
+	vnl_matrix <double> tableToMatrix(vtkSmartPointer<vtkTable> table,std::vector< std::pair<int,int> > id_list);
 	vnl_matrix <double> tableToMatrix_1(vtkSmartPointer<vtkTable> table);
 	vnl_matrix <double> Normalize_Feature_Matrix(vnl_matrix<double> feats);
 	vnl_matrix <double> Normalize_Feature_Matrix_1(vnl_matrix<double> feats, vnl_vector<double> vector_1, vnl_vector<double> vector_2);
